Feat: move transforms to dataset, n2v to lightning
