+++ src/careamics/dataset/dataset.py
@@ -13,10 +13,10 @@ from careamics.config.data.data_config import (
     Mode,
     WholePatchingConfig,
 )
+from careamics.dataset.transforms.compose import Compose
 from careamics.lightning.lightning_modules.constraints import (
     ModelConstraints,
 )
-from careamics.transforms.compose import Compose
 
 from .image_stack import GenericImageStack, ZarrImageStack
 from .normalization import create_normalization
