+++ src/careamics/dataset/transforms/__init__.py
@@ -1,13 +1,11 @@
 """Transforms that are used to augment the data."""
 
 __all__ = [
-    "N2VManipulate",
     "Transform",
     "XYFlip",
     "XYRandomRotate90",
 ]
 
-from .n2v_manipulate import N2VManipulate
 from .transform import Transform
 from .xy_flip import XYFlip
 from .xy_random_rotate90 import XYRandomRotate90
