+++ src/careamics/lightning/lightning_modules/n2v_utils/pixel_manipulation.py
