+++ src/careamics/lightning/lightning_modules/n2v_utils/struct_mask_parameters.py
