+++ tests/dataset/transforms/test_xy_flip.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pytest
 
-from careamics.transforms import XYFlip
+from careamics.dataset.transforms import XYFlip
 
 
 def test_p_wrong_values():
