+++ tests/lightning/n2v_utils/test_pixel_manipulation.py
@@ -5,7 +5,7 @@ import pytest
 import torch
 from torch.nn.functional import pad
 
-from careamics.transforms.pixel_manipulation import (
+from careamics.lightning.lightning_modules.n2v_utils.pixel_manipulation import (
     _apply_struct_mask_torch,
     _create_center_pixel_mask,
     _create_struct_mask,
@@ -14,7 +14,9 @@ from careamics.transforms.pixel_manipulation import (
     median_manipulate,
     uniform_manipulate,
 )
-from careamics.transforms.struct_mask_parameters import StructMaskParameters
+from careamics.lightning.lightning_modules.n2v_utils.struct_mask_parameters import (
+    StructMaskParameters,
+)
 
 
 @pytest.mark.parametrize(
